+++ crates/store/re_protos/src/lib.rs
@@ -5,8 +5,6 @@
 //! to use gRPC types in the rerun viewer codebase. That's why we implement all the
 //! necessary conversion code (in the form of `From` and `TryFrom` traits) in this crate.
 
-mod protobuf_conversions;
-
 pub mod external {
     pub use prost;
 }
