--- crates/store/re_protos/src/protobuf_conversions/mod.rs
@@ -1,1 +1,0 @@
-mod arrow;
